Data Engineer/object-storage/requirements.txt
boto3
python-dotenv
SQLAlchemy
pandas